Reworks SaibaneCode table to resolve primary key constraint violations

+++ Solidus.Suica/Data/Entities/EkiData/Company.cs
@@ -10,9 +10,9 @@
         public string? OfficialName { get; set; }
         public string? ShortName { get; set; }
         public string? Website { get; set; }
-        public short? CompanyTypeId { get; set; }
+        public byte? CompanyTypeId { get; set; }
         public CompanyType? Type { get; set; }
-        public short? StatusId { get; set; }
+        public byte? StatusId { get; set; }
         public Status? Status { get; set; }
     }
 }
