+++ Solidus.Suica/Data/Entities/EkiData/LineType.cs
@@ -2,7 +2,7 @@
 {
     public class LineType
     {
-        public short Id { get; set; }
+        public byte Id { get; set; }
         public string Name_JA { get; set; }
         public string Name_EN { get; set; }
     }
