+++ Solidus.Suica/Data/Entities/EkiData/Station.cs
@@ -9,7 +9,7 @@ namespace Solidus.SuicaTools.Data.Entities.EkiData
         public string StationName { get; set; }
         public int LineCode { get; set; }
         public Line Line { get; set; }
-        public short? PrefectureId { get; set; } 
+        public byte? PrefectureId { get; set; } 
         public Prefecture? Prefecture { get; set; }
         public string? PostalCode { get; set; }
         public string? Address { get; set; }
@@ -17,7 +17,7 @@ namespace Solidus.SuicaTools.Data.Entities.EkiData
         public decimal? Latitude { get; set; }
         public DateTime? OpeningDate { get; set; }
         public DateTime? ClosingDate { get; set; }
-        public short? StatusId { get; set; }
+        public byte? StatusId { get; set; }
         public Status? Status { get; set; }
     }
 }
